Implement inflating presentation class from XML.
The Presentation class no longer takes an array of slides in the constructor;
instead, there's a setSlides() method which will be invoked either by
subclasses or the inflater.

MainActivity loads the xml/main_presentation resource instead of overriding
the xml/presentation resource -- this is temporary to make it easier to
switch between the default (sample) presentation and the regular presentation.

Issue #14.